bot-agenda/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
APScheduler==3.11.2
certifi==2026.2.25
charset-normalizer==3.4.7
colorama==0.4.6
defusedxml==0.7.1
distro==1.9.0
elevenlabs==2.44.0
fonttools==4.62.1
fpdf2==2.8.7
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.12
jiter==0.14.0
openai==2.32.0
pillow==12.2.0
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
python-telegram-bot==22.7
pytz==2026.1.post1
requests==2.33.1
sniffio==1.3.1
SQLAlchemy==2.0.49
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
websockets==16.0
